+++ common/error.go
@@ -1,0 +1,11 @@
+package common
+
+import (
+	"github.com/juju/errgo"
+)
+
+var (
+	maskAny = errgo.MaskFunc(errgo.Any)
+)
+
+var invalidArgumentsError = errgo.Newf("invalid arguments")
