+++ fleet/fleet.go
@@ -9,11 +9,12 @@ import (
 	"net/http"
 	"net/url"
 	"regexp"
-	"strings"
 
 	"github.com/coreos/fleet/client"
 	"github.com/coreos/fleet/schema"
 	"github.com/coreos/fleet/unit"
+
+	"github.com/giantswarm/formica/common"
 )
 
 const (
@@ -78,110 +79,6 @@ type UnitStatus struct {
 
 	// Slice represents the slice expression. E.g. @1, or @foo, or @5., etc..
 	Slice string
-}
-
-type UnitStatusList []UnitStatus
-
-func (usl UnitStatusList) Group() ([]UnitStatus, error) {
-	matchers := map[string]struct{}{}
-	newList := []UnitStatus{}
-
-	for _, us := range usl {
-		// Group unit status
-		grouped, suffix, err := groupUnitStatus(usl, us)
-		if err != nil {
-			return nil, maskAny(err)
-		}
-
-		// Prevent doubled aggregation.
-		if _, ok := matchers[suffix]; ok {
-			continue
-		}
-		matchers[suffix] = struct{}{}
-
-		// Aggregate.
-		if allStatesEqual(grouped) {
-			newStatus := grouped[0]
-			newStatus.Name = "*"
-			newList = append(newList, newStatus)
-		} else {
-			newList = append(newList, grouped...)
-		}
-	}
-
-	return newList, nil
-}
-
-func groupUnitStatus(usl []UnitStatus, groupMember UnitStatus) ([]UnitStatus, string, error) {
-	ID, err := sliceID(groupMember.Name)
-	if err != nil {
-		return nil, "", maskAny(invalidUnitStatusError)
-	}
-
-	newList := []UnitStatus{}
-	for _, us := range usl {
-		exp := extExp.ReplaceAllString(us.Name, "")
-		if !strings.HasSuffix(exp, ID) {
-			continue
-		}
-
-		newList = append(newList, us)
-	}
-
-	return newList, ID, nil
-}
-
-// extExp matches unit file extensions.
-//
-//   app@1.service  =>  .service
-//   app@1.mount    =>  .mount
-//   app.service    =>  .service
-//   app.mount      =>  .mount
-//
-var extExp = regexp.MustCompile(`(?m)\.[a-z]*$`)
-
-func sliceID(name string) (string, error) {
-	suffix, err := sliceSuffix(name)
-	if err != nil {
-		return "", maskAny(err)
-	}
-	ID := extExp.ReplaceAllString(suffix, "")
-
-	return ID, nil
-}
-
-var groupExp = regexp.MustCompile("@(.*)")
-
-func sliceSuffix(name string) (string, error) {
-	found := groupExp.FindAllString(name, -1)
-	if len(found) == 0 {
-		return "", nil
-	} else if len(found) > 1 {
-		return "", maskAny(invalidUnitStatusError)
-	}
-	return found[0], nil
-}
-
-func allStatesEqual(usl []UnitStatus) bool {
-	for _, us1 := range usl {
-		for _, us2 := range usl {
-			if us1.Current != us2.Current {
-				return false
-			}
-			if us1.Desired != us2.Desired {
-				return false
-			}
-			for _, m1 := range us1.Machine {
-				for _, m2 := range us2.Machine {
-					if m1.SystemdActive != m2.SystemdActive {
-						return false
-					}
-				}
-			}
-		}
-	}
-
-	return true
 }
 
 // Fleet defines the interface a fleet client needs to implement to provide
@@ -398,7 +295,7 @@ func (f fleet) createOurStatusList(foundFleetUnits []*schema.Unit, foundFleetUni
 	ourStatusList := []UnitStatus{}
 
 	for _, ffu := range foundFleetUnits {
-		ID, err := sliceID(ffu.Name)
+		ID, err := common.SliceID(ffu.Name)
 		if err != nil {
 			return nil, maskAny(invalidUnitStatusError)
 		}
